requirements.txt
fastapi
uvicorn
sse-starlette
supabase
anthropic
openai
python-dotenv
pydantic
